requirements.txt
matplotlib
seaborn
numpy
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pypng>=0.20220715.0
pillow-avif-plugin>=1.3.1
qoi>=0.2.0
pytest>=7.0
pytest-cov>=3.0
pytest-mock>=3.0
psutil>=5.9